% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-description.r
\docType{data}
\name{closures}
\alias{closures}
\title{Estuary Closures}
\format{A data frame with 3 variables}
\usage{
data(closures)
}
\description{
Estuary closures identified from water level and photographic data. 
Structured as:
\enumerate{
  \item \code{id} (arbitary) closure id
  \item \code{initiation} date of closure initiation
  \item \code{breach} date of mouth breach
}
}
\keyword{datasets}

