% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-description.r
\docType{data}
\name{ctd}
\alias{ctd}
\title{CTD Cast Data}
\format{A data frame with 14 variables}
\usage{
data(ctd)
}
\description{
CTD cast data collected by the Largier Lab and processed using 
\code{rremat}. Structured as:
\enumerate{
  \item \code{date} cast date
  \item \code{dist} longitudinal distance from mouth (km)
  \item \code{surfelev} water surface elevation (m NAVD29)
  \item \code{elev} elevation (m NAVD29)
  \item \code{depth} cast depth (m)
  \item \code{ta} temperature (degrees C)
  \item \code{sa} salinity (PSU)
  \item \code{da}       
  \item \code{oa} dissolved oxygen (mg/L)        
  \item \code{sat} dissolved oxygen percent saturation (%)
  \item \code{fl} fluorescence (ug/L))
  \item \code{bt} beam transmission (%)
  \item \code{par} photosynthetically-active radiation (umol^-1 m^-2)
  \item \code{ph} pH
}
}
\seealso{
ctdmeta grids habgrids
}
\keyword{datasets}

