% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-description.r
\docType{data}
\name{ctdmeta}
\alias{ctdmeta}
\title{CTD Transect Meta Data}
\format{A data frame with 3 variables}
\usage{
data(ctdmeta)
}
\description{
Meta data for CTD transects. Structured as:
\itemize{
  \item \code{start} timestamp of transect commencement
  \item \code{end} timestamp of transect completion
  \item \code{note} additional note on e.g. mouth state
}
}
\details{
Transect meta data was collected from the Russian River Estuary 
  Circulation and Water Quality Data Reports (2011, 2012, 2013, 2014) 
  submitted to the Sonoma County Water Agency.
}
\keyword{datasets}

