% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-description.r
\docType{data}
\name{grids}
\alias{grids}
\title{Interpolated CTD Grids}
\format{A data frame with 6 variables}
\usage{
data(grids)
}
\description{
Interpolated CTD grids constructed using natural neighbor interpolation in 
Matlab. Structured as:
\itemize{
  \item \code{sa} salinity (PSU)
  \item \code{ta} temperature (degrees C)
  \item \code{oa} dissolved oxygen (mg/L)        
  \item \code{dist} longitudinal distance from mouth (km)
  \item \code{elev} elevation (m NAVD29)
  \item \code{date} cast date
}
}
\seealso{
ctdmeta ctd
}
\keyword{datasets}

