% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-description.r
\docType{data}
\name{habgrids}
\alias{habgrids}
\title{Habitat Data}
\format{A data frame with 6 variables}
\usage{
data(habgrids)
}
\description{
Habitat evaluation of interpolated CTD transects. Same structure and data
as \code{grids}, but with the following additional columns:
}
\seealso{
grids
}
\keyword{datasets}

