% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-description.r
\docType{package}
\name{habitatblueprint-package}
\alias{habitatblueprint-package}
\title{Habitat Blueprint Browser}
\description{
This packages contains the Habitat Blueprint Browser shiny app. Data used 
by the app was generated using the \code{rremat} package.
}

