% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-description.r
\name{HabitatBrowser}
\alias{HabitatBrowser}
\title{Habitat Blueprint Browser}
\usage{
HabitatBrowser()
}
\description{
Start the Habitat Blueprint Browser.
}
\examples{
\dontrun{
  HabitatBrowser()
}
}

