% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-description.r
\docType{data}
\name{inflows}
\alias{inflows}
\title{River Flow Data}
\format{A data frame with 3 variables}
\usage{
data(inflows)
}
\description{
River flow data managed by USGS. Structured as:
\itemize{
  \item \code{site} gauge location. Austin Creek near Cazadero and Russian 
        River near Guerneville are included.
  \item \code{datetime} timestamp (PST/PDT)
  \item \code{flow} flow rate (cfs)
}
}
\keyword{datasets}

