% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-description.r
\docType{data}
\name{tides}
\alias{tides}
\title{Tide Height Data}
\format{A data frame with 3 variables}
\usage{
data(tides)
}
\description{
Tide height data managed by NOAA. Structured as:
\itemize{
  \item \code{datetime} timestamp (UTC)
  \item \code{height} Tide height above MLLW (m)
  \item \code{sigma}
}
}
\keyword{datasets}

