% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-description.r
\docType{data}
\name{wll}
\alias{wll}
\title{Water Level Data}
\format{A data frame with 3 variables}
\usage{
data(wll)
}
\description{
Pressure gauge water Level data collected by the Largier Lab and processed 
using \code{rremat}. Structured as:
\itemize{
  \item \code{site} gauge location
  \item \code{mtime} timestamp (UTC)
  \item \code{depth} water depth (m)
}
}
\keyword{datasets}

